+++ core/src/feature/metal/float_adm_metal.mm
@@ -216,7 +216,7 @@ static const VmafOption options[] = {
      .default_val = {.i = DEFAULT_ADM_CSF_MODE},
      .min = 0,
      .max = 9,
-     .flags = VMAF_OPT_FLAG_FEATURE_PARAM},
+     .flags = VMAF_OPT_FLAG_FEATURE_PARAM | VMAF_OPT_FLAG_DEFAULT_ONLY},
     {.name = "adm_csf_scale",
      .alias = "scf",
      .help = "scale factor for the CSF",