+++ core/src/feature/sycl/float_vif_sycl.cpp
@@ -665,7 +665,7 @@ static const VmafOption options_float_vif_sycl[] = {
      .default_val = {.d = 1.0},
      .min = 0.1,
      .max = 4.0,
-     .flags = VMAF_OPT_FLAG_FEATURE_PARAM},
+     .flags = VMAF_OPT_FLAG_FEATURE_PARAM | VMAF_OPT_FLAG_DEFAULT_ONLY},
     {.name = "vif_sigma_nsq",
      .help = "neural noise variance",
      .alias = "snsq",